Rename class and package
