+++ deskover-ecommerce-website/src/main/java/com/deskover/util/ValidationUtils.java
@@ -7,7 +7,7 @@ import org.springframework.validation.BindingResult;
 
 
 
-public class ValidationUtil {
+public class ValidationUtils {
 
     /**
      * Convert BindingResult to List<ValidationError>
